requirements.txt
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
dnspython==2.2.0
Flask==2.0.2
Flask-Cors==3.0.10
google-auth==2.6.0
google-auth-oauthlib==0.5.0
gspread==5.2.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauth2client==4.1.3
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pymongo==4.0.1
pyparsing==3.0.7
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
